api/requirements.txt
flask
flask-cors
pillow
reportlab
pillow-heif